Test review Core J-S, #259 (#914)
* Formatting and remove extraneous comment in Search/Payloads, #259

* Array formatting in Search/Similarities to match upstream

* Search/Spans formatting cleanup, use singletons for parameterless/captureless anonymous classes

* Code cleanup in Search/B-TestB tests

* Search/TestC code cleanup and allocation improvements

* Search/TestD-TestE code cleanup

* Test review TestF-TestL, #259

* Test review for rest of Search, #259

* Test review of Store, #259

* Finish J-S test cleanup

* Fix .NET FX build failure due to missing import

* Use CompareToOrdinal instead of string.Compare

* Change Array.Empty<T> to Arrays.Empty<T> until #916 is done in a separate PR

* Null-safe disposal of resources in TestSpansAdvanced

* Make TestSpansAdvanced.AssertHits static to match upstream Java code

* Use Arrays.Empty<T> in TestBoolean2 instead of Array.Empty<T>

* Revert static instance of MockScorer in TestCachingCollector

* Revert static instances of NoOpCollector in TestCachingCollector

* Fix seealso cref to CheckHits.CheckNoMatchExplanations

* Fix ticks math in RandomGen.LuceneDate

* Fix XML doc comment paragraph tag in TestFieldCacheRangeFilter

* Revert static instance of AnalyzerAnonymousClass in TestPhraseQuery

* Comment out TestScorerPerf.terms to prevent unused warning

* Remove use of Convert.ToInt32 which boxes in CheckHits

* Specify culture for Convert.ToInt32

* Fix spelling of TransactionalThreadInterrupt class

* Add back redundant override of TestRegexps in case some test runners don't report failures correctly

* Remove unused ConcurrentDictionaryWrapper type

+++ src/Lucene.Net.Tests/Search/TestFieldCacheRangeFilter.cs
@@ -36,7 +36,7 @@ namespace Lucene.Net.Search
     /// <summary>
     /// A basic 'positive' Unit test class for the FieldCacheRangeFilter class.
     ///
-    /// <p>
+    /// <para />
     /// NOTE: at the moment, this class only tests for 'positive' results,
     /// it does not verify the results to ensure there are no 'false positives',
     /// nor does it adequately test 'negative' results.  It also does not test
@@ -47,8 +47,8 @@ namespace Lucene.Net.Search
     {
         /// <summary>
         /// LUCENENET specific. Ensure we have an infostream attached to the default FieldCache
-        /// when running the tests. In Java, this was done in the Core.Search.TestFieldCache.TestInfoStream() 
-        /// method (which polluted the state of these tests), but we need to make the tests self-contained 
+        /// when running the tests. In Java, this was done in the Core.Search.TestFieldCache.TestInfoStream()
+        /// method (which polluted the state of these tests), but we need to make the tests self-contained
         /// so they can be run correctly regardless of order. Not setting the InfoStream skips an execution
         /// path within these tests, so we should do it to make sure we test all of the code.
         /// </summary>
@@ -501,7 +501,7 @@ namespace Lucene.Net.Search
 
             int numDocs = reader.NumDocs;
             float minIdO = Convert.ToSingle(minId + .5f);
-            float medIdO = Convert.ToSingle((float)minIdO + ((maxId - minId)) / 2.0f);
+            float medIdO = Convert.ToSingle(minIdO + ((maxId - minId)) / 2.0f);
 
             ScoreDoc[] result;
             Query q = new TermQuery(new Term("body", "body"));
